+++ developer/src/kmc-kmn/src/compiler/compiler.ts
@@ -72,11 +72,11 @@ class WasmWrapper {
     this.parseUnicodeSet = this.Module.cwrap('kmcmp_Wasm_ParseUnicodeSet', 'number', ['string', 'number', 'number']);
     this.setCompilerOptions = this.Module.cwrap('kmcmp_Wasm_SetCompilerOptions', 'boolean', ['number']);
 
-    if (this.parseUnicodeSet == undefined
-      || this.setCompilerOptions == undefined
-      || this.compileKeyboardFile == undefined) {
+    if (this.parseUnicodeSet === undefined
+      || this.setCompilerOptions === undefined
+      || this.compileKeyboardFile === undefined) {
         throw Error(`some wasm functions did not load properly.`);
-      }
+    }
   }
 
   /**
@@ -88,7 +88,7 @@ class WasmWrapper {
   }
 };
 
-export class Compiler implements UnicodeSetParser {
+export class KmnCompiler implements UnicodeSetParser {
   callbackName: string;
   callbacks: CompilerCallbacks;
   wasm: WasmWrapper;
@@ -99,9 +99,7 @@ export class Compiler implements UnicodeSetParser {
   }
 
   public async init(callbacks: CompilerCallbacks): Promise<boolean> {
-    if(!this.callbacks) {
-      this.callbacks = callbacks;
-    }
+    this.callbacks = callbacks;
     if(!this.wasm) {
       try {
         this.wasm = await WasmWrapper.load();
@@ -110,14 +108,14 @@ export class Compiler implements UnicodeSetParser {
         return false;
       }
     }
-    return this.verifyInitted();
+    return this.verifyInitialized();
   }
 
   /**
    * Verify that wasm is spun up OK.
    * @returns true if OK
    */
-  public verifyInitted() : boolean {
+  public verifyInitialized() : boolean {
     if(!this.callbacks) {
       // Can't report a message here.
       throw Error('Must call Compiler.init(callbacks) before proceeding');
@@ -130,7 +128,7 @@ export class Compiler implements UnicodeSetParser {
   }
 
   public run(infile: string, outfile: string, options?: CompilerOptions): boolean {
-    if(!this.verifyInitted()) {
+    if(!this.verifyInitialized()) {
       return false;
     }
 
@@ -172,7 +170,7 @@ export class Compiler implements UnicodeSetParser {
    * @returns UnicodeSet accessor object, or null on failure
    */
   public parseUnicodeSet(pattern: string, bufferSize: number) : UnicodeSet | null {
-    if(!this.verifyInitted()) {
+    if(!this.verifyInitialized()) {
       return null;
     }