+++ developer/src/kmc-ldml/src/compiler/compiler.ts
@@ -1,5 +1,5 @@
 import { LDMLKeyboardXMLSourceFileReader, LDMLKeyboard, KMXPlus, CompilerCallbacks, LDMLKeyboardTestDataXMLSourceFile } from '@keymanapp/common-types';
-import CompilerOptions from './compiler-options.js';
+import { CompilerOptions } from './compiler-options.js';
 import { CompilerMessages } from './messages.js';
 import { BkspCompiler, TranCompiler } from './tran.js';
 import { DispCompiler } from './disp.js';
@@ -31,7 +31,7 @@ const SECTION_COMPILERS = [
   VkeyCompiler,
 ];
 
-export default class Compiler {
+export class LdmlKeyboardCompiler {
   private readonly callbacks: CompilerCallbacks;
   // private readonly options: CompilerOptions; // not currently used